zsh: Source .shtools configuration
For some shell tools, such as pyenv, they require making changes to
your shell config. To keep my shell config simple, I will *not* have
the config in .zshrc, but have them in a file ~/.shtools. And source
them from .zshrc

I might rename the file in the future to basically capture all
configuration that is external, and maybe check that in if needed